\name{functionClassicA}
\Rdversion{1.1}
\alias{functionClassicA}
\title{functionClassicA}
\description{A very nice function}

\usage{
functionClassicA(age)
}
\arguments{
  \item{age}{[numeric] The age of the patient}
}
\details{Double the age}

\value{An age [numeric]}

\author{Christophe Genolini <genolini@u-paris10.fr>}

\references{
  Book: "Petit Manuel de Programmation Orientee Objet sous R"
}

\examples{
myAge <- 35
functionClassicA(myAge)
}

\keyword{documentation}

