\name{publicA-methods}
\docType{methods}
\alias{publicA,ClassV-method}
\title{publicA for ClassV}
\description{
  publicA is a great function.
}
\section{Methods}{
  \describe{
    \item{object = "ClassV"}{publicA on object ClassV}
  }
}
\author{Christophe Genolini <genolini@u-paris10.fr>}

\references{
  Book: "Petit Manuel de Programmation Orientee Objet sous R"
}

\examples{
  vvv <- new("ClassV",v1=2,v2=-4)
  publicA(vvv)
}
\keyword{documentation}

