\name{publicA-methods}
\docType{methods}
\alias{publicA,ClassW-method}
\title{publicA for ClassW}
\description{
  publicA is a great function.
}
\section{Methods}{
  \describe{
    \item{object = "ClassW"}{publicA on object ClassW}
  }
}

\author{Christophe Genolini <genolini@u-paris10.fr>}

\references{
  Book: "Petit Manuel de Programmation Orientee Objet sous R"
}

\examples{
  www <- classW(w1="A",u1=2,u2=-4)
  publicA(www)
}

\keyword{documentation}


