\name{publicB}
\Rdversion{1.1}
\alias{publicB}
\alias{publicB-methods}
\alias{publicB,ClassV,ClassW-method}

\title{publicB}
\description{publicB is a great function too}

\usage{
publicB(objectV, objectW)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{objectV}{Object of classV}
  \item{objectW}{Object of classV}
}
\details{
  publicB is a great function
}
\value{
  A numeric
}

\author{Christophe Genolini <genolini@u-paris10.fr>}

\references{
  Book: "Petit Manuel de Programmation Orientee Objet sous R"
}

\examples{
vvv <- new("ClassV",v1=1,v2=2)
www <- classW(w1="Z",u1=2,u2=3)
publicB(vvv,www)
}

\keyword{documentation}

